// client/src/app/movies/models/movie.ts
export interface Movie {
  poster_path: string;
  adult: false;
  overview: string;
  relase_date: string;
  genre_ids: number[];
  id: number;
  original_title: string;
  original_language: string;
  title: string;
  backdrop_path: string;
  popularity: number;
  vote_count: number;
  video: boolean;
  vote_average: number;
}